+++ imod/mf6/chd.py
@@ -1,11 +1,9 @@
-from typing import Optional, Tuple
-
 import numpy as np
 
 from imod.logging import init_log_decorator
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
-from imod.mf6.utilities.regrid import RegridderType
+from imod.mf6.regrid.regrid_schemes import ConstantHeadRegridMethod
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
 from imod.schemata import (
     AllInsideNoDataSchema,
@@ -108,14 +106,7 @@ class ConstantHead(BoundaryCondition, IRegridPackage):
     _keyword_map = {}
     _auxiliary_data = {"concentration": "species"}
     _template = BoundaryCondition._initialize_template(_pkg_id)
-
-    _regrid_method = {
-        "head": (
-            RegridderType.OVERLAP,
-            "mean",
-        ),  # TODO: should be set to barycentric once supported
-        "concentration": (RegridderType.OVERLAP, "mean"),
-    }
+    _regrid_method = ConstantHeadRegridMethod()
 
     @init_log_decorator()
     def __init__(
@@ -141,6 +132,7 @@ class ConstantHead(BoundaryCondition, IRegridPackage):
             "repeat_stress": repeat_stress,
         }
         super().__init__(dict_dataset)
+
         self._validate_init_schemata(validate)
 
     def _validate(self, schemata, **kwargs):
@@ -149,6 +141,3 @@ class ConstantHead(BoundaryCondition, IRegridPackage):
         errors = super()._validate(schemata, **kwargs)
 
         return errors
-
-    def get_regrid_methods(self) -> Optional[dict[str, Tuple[RegridderType, str]]]:
-        return self._regrid_method