+++ imod/mf6/cnc.py
@@ -1,11 +1,8 @@
-from typing import Optional, Tuple
-
 import numpy as np
 
 from imod.logging import init_log_decorator
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
-from imod.mf6.utilities.regridding_types import RegridderType
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA
 from imod.schemata import (
     AllInsideNoDataSchema,
@@ -53,8 +50,6 @@ class ConstantConcentration(BoundaryCondition, IRegridPackage):
     _period_data = ("concentration",)
     _template = BoundaryCondition._initialize_template(_pkg_id)
 
-    _regrid_method: dict[str, tuple[RegridderType, str]] = {}
-
     _init_schemata = {
         "concentration": [
             DTypeSchema(np.floating),
@@ -91,6 +86,3 @@ class ConstantConcentration(BoundaryCondition, IRegridPackage):
         }
         super().__init__(dict_dataset)
         self._validate_init_schemata(validate)
-
-    def get_regrid_methods(self) -> Optional[dict[str, Tuple[RegridderType, str]]]:
-        return self._regrid_method