+++ imod/mf6/package.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import abc
 import pathlib
 from collections import defaultdict
+from copy import deepcopy
 from typing import Any, Mapping, Optional, Tuple, Union
 
 import cftime
@@ -22,10 +23,10 @@ from imod.mf6.pkgbase import (
     TRANSPORT_PACKAGES,
     PackageBase,
 )
+from imod.mf6.regrid.regrid_schemes import EmptyRegridMethod, RegridMethodType
 from imod.mf6.utilities.mask import mask_package
 from imod.mf6.utilities.package import _is_valid
 from imod.mf6.utilities.regrid import (
-    RegridderType,
     RegridderWeightsCache,
     _regrid_like,
 )
@@ -58,6 +59,7 @@ class Package(PackageBase, IPackage, abc.ABC):
     _init_schemata: dict[str, list[SchemaType] | Tuple[SchemaType, ...]] = {}
     _write_schemata: dict[str, list[SchemaType] | Tuple[SchemaType, ...]] = {}
     _keyword_map: dict[str, str] = {}
+    _regrid_method: RegridMethodType = EmptyRegridMethod()
 
     def __init__(self, allargs: Mapping[str, GridDataArray | float | int | bool | str]):
         super().__init__(allargs)
@@ -540,35 +542,38 @@ class Package(PackageBase, IPackage, abc.ABC):
         self,
         target_grid: GridDataArray,
         regrid_context: RegridderWeightsCache,
-        regridder_types: Optional[dict[str, Tuple[RegridderType, str]]] = None,
+        regridder_types: Optional[RegridMethodType] = None,
     ) -> "Package":
         """
-        Creates a package of the same type as this package, based on another discretization.
-        It regrids all the arrays in this package to the desired discretization, and leaves the options
-        unmodified. At the moment only regridding to a different planar grid is supported, meaning
-        ``target_grid`` has different ``"x"`` and ``"y"`` or different ``cell2d`` coords.
+        Creates a package of the same type as this package, based on another
+        discretization. It regrids all the arrays in this package to the desired
+        discretization, and leaves the options unmodified. At the moment only
+        regridding to a different planar grid is supported, meaning
+        ``target_grid`` has different ``"x"`` and ``"y"`` or different
+        ``cell2d`` coords.
 
-        The regridding methods can be specified in the _regrid_method attribute of the package. These are the defaults
-        that specify how each array should be regridded. These defaults can be overridden using the input
-        parameters of this function.
+        The default regridding methods are specified in the ``_regrid_method``
+        attribute of the package. These defaults can be overridden using the
+        input parameters of this function.
 
         Examples
         --------
         To regrid the npf package with a non-default method for the k-field, call regrid_like with these arguments:
 
-        >>> new_npf = npf.regrid_like(like, {"k": (imod.RegridderType.OVERLAP, "mean")})
+        >>> regridder_types = imod.mf6.regrid.NodePropertyFlowRegridMethod(k=(imod.RegridderType.OVERLAP, "mean"))
+        >>> new_npf = npf.regrid_like(like,  RegridderWeightsCache, regridder_types)
 
 
         Parameters
         ----------
         target_grid: xr.DataArray or xu.UgridDataArray
-            a grid defined over the same discretization as the one we want to regrid the package to
-        regridder_types: dict(str->(regridder type,str))
-           dictionary mapping arraynames (str) to a tuple of regrid type (a specialization class of BaseRegridder) and function name (str)
-            this dictionary can be used to override the default mapping method.
-        regrid_context: Optional RegridderWeightsCache
+            a grid defined over the same discretization as the one we want to regrid the package to.
+        regrid_context: RegridderWeightsCache, optional
             stores regridder weights for different regridders. Can be used to speed up regridding,
             if the same regridders are used several times for regridding different arrays.
+        regridder_types: RegridMethodType, optional
+            dictionary mapping arraynames (str) to a tuple of regrid type (a specialization class of BaseRegridder) and function name (str)
+            this dictionary can be used to override the default mapping method.
 
         Returns
         -------
@@ -631,3 +636,6 @@ class Package(PackageBase, IPackage, abc.ABC):
 
     def is_clipping_supported(self) -> bool:
         return True
+
+    def get_regrid_methods(self) -> RegridMethodType:
+        return deepcopy(self._regrid_method)