+++ imod/mf6/rch.py
@@ -1,5 +1,5 @@
 from copy import deepcopy
-from typing import Optional, Tuple
+from typing import Optional
 
 import numpy as np
 
@@ -7,11 +7,9 @@ from imod.logging import init_log_decorator
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.dis import StructuredDiscretization
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
-from imod.mf6.utilities.regrid import (
-    RegridderType,
-    RegridderWeightsCache,
-    _regrid_package_data,
-)
+from imod.mf6.regrid.regrid_schemes import RechargeRegridMethod
+from imod.mf6.utilities.regrid import RegridderWeightsCache, _regrid_package_data
+from imod.mf6.utilities.regridding_types import RegridderType
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
 from imod.prepare.topsystem.allocation import ALLOCATION_OPTION, allocate_rch_cells
 from imod.schemata import (
@@ -121,11 +119,7 @@ class Recharge(BoundaryCondition, IRegridPackage):
 
     _template = BoundaryCondition._initialize_template(_pkg_id)
     _auxiliary_data = {"concentration": "species"}
-
-    _regrid_method = {
-        "rate": (RegridderType.OVERLAP, "mean"),
-        "concentration": (RegridderType.OVERLAP, "mean"),
-    }
+    _regrid_method = RechargeRegridMethod()
 
     @init_log_decorator()
     def __init__(
@@ -161,9 +155,6 @@ class Recharge(BoundaryCondition, IRegridPackage):
         errors = super()._validate(schemata, **kwargs)
 
         return errors
-
-    def get_regrid_methods(self) -> Optional[dict[str, Tuple[RegridderType, str]]]:
-        return self._regrid_method
 
     @classmethod
     def from_imod5_data(