+++ imod/mf6/riv.py
@@ -1,6 +1,4 @@
 from copy import deepcopy
-from typing import Optional, Tuple
-
 import numpy as np
 import xarray as xr
 
@@ -9,7 +7,7 @@ from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.dis import StructuredDiscretization
 from imod.mf6.drn import Drainage
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
-from imod.mf6.npf import NodePropertyFlow
+from imod.mf6.regrid.regrid_schemes import RiverRegridMethod
 from imod.mf6.utilities.regrid import (
     RegridderType,
     RegridderWeightsCache,
@@ -141,13 +139,7 @@ class River(BoundaryCondition, IRegridPackage):
 
     _template = BoundaryCondition._initialize_template(_pkg_id)
     _auxiliary_data = {"concentration": "species"}
-
-    _regrid_method = {
-        "stage": (RegridderType.OVERLAP, "mean"),
-        "conductance": (RegridderType.RELATIVEOVERLAP, "conductance"),
-        "bottom_elevation": (RegridderType.OVERLAP, "mean"),
-        "concentration": (RegridderType.OVERLAP, "mean"),
-    }
+    _regrid_method = RiverRegridMethod()
 
     @init_log_decorator()
     def __init__(
@@ -177,7 +169,6 @@ class River(BoundaryCondition, IRegridPackage):
             "repeat_stress": repeat_stress,
         }
         super().__init__(dict_dataset)
-
         self._validate_init_schemata(validate)
 
     def _validate(self, schemata, **kwargs):
@@ -188,8 +179,6 @@ class River(BoundaryCondition, IRegridPackage):
 
         return errors
 
-    def get_regrid_methods(self) -> Optional[dict[str, Tuple[RegridderType, str]]]:
-        return self._regrid_method
 
     @classmethod
     def from_imod5_data(